requirements.txt
# Core dependencies
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0

# MetaTrader5 integration
MetaTrader5==5.0.45

# Async support
aiohttp==3.9.1
asyncio-mqtt==0.16.1

# Configuration and validation
pydantic==2.5.2
pydantic-settings==2.1.0

# Logging
structlog==23.2.0

# Database
asyncpg==0.29.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development
black==23.12.1
isort==5.13.2
flake8==6.1.0
mypy==1.8.0

# Monitoring
prometheus-client==0.19.0 